:recycle: BTS-300 refactor: http 계층 명확화 및 네이밍 개선
- http.transport.ts: Axios 인스턴스 (Infra)로 역할 분리
- http.client.ts: 도메인에서 사용하는 HttpClient 추상화(Port+Adapter)로 정의
- http.port.ts: HttpClient 타입 계약 정의
- src/http/index.ts 에서 export 관리
