+++ src/lib/http/http.port.ts
@@ -1,0 +1,3 @@
+import { authApi } from '@/lib/http';
+
+export type HttpClient = typeof authApi;
